backend/requirements.txt
# FastAPI and server
fastapi==0.123.9
uvicorn==0.38.0
python-dotenv==1.2.1

# LLM and AI
openai==2.9.0
openai-agents==0.6.2
tiktoken==0.12.0

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.2

# Vector store
qdrant-client==1.16.1

# Pydantic
pydantic==2.12.5
pydantic-settings==2.12.0

# Testing
pytest==9.0.1
pytest-asyncio==1.3.0